requirements.txt
streamlit
numpy
faiss-cpu==1.7.3
ftfy
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
--extra-index-url https://download.pytorch.org/whl/cu117
torchvision==0.14.1+cu117
--extra-index-url https://download.pytorch.org/whl/cu117
torchaudio==0.13.1 
regex
tqdm
git+https://github.com/openai/CLIP.git
